move MemoryExperimentParts
