+++ src/qldpc/circuits/bookkeeping.py
@@ -22,7 +22,6 @@ import copy
 import dataclasses
 import itertools
 from collections.abc import Hashable, ItemsView, Iterable, Iterator, Mapping, Sequence
-from typing import NamedTuple
 
 import numpy as np
 import stim
@@ -274,12 +273,3 @@ class DetectorRecord(Record):
                 if other_key != key
             }
         )
-
-
-class MemoryExperimentParts(NamedTuple):
-    initialization: stim.Circuit
-    qec_cycle: stim.Circuit
-    readout: stim.Circuit
-    measurement_record: MeasurementRecord
-    detector_record: DetectorRecord
-    qubit_ids: QubitIDs
